reduced code redundancy
